% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{effective_sample_size}
\alias{effective_sample_size}
\title{Calculate Effective Sample Size}
\usage{
effective_sample_size(m, a0)
}
\arguments{
\item{m}{Sample size of historical data}

\item{a0}{Discounting parameter}
}
\value{
Effective sample size (numeric)
}
\description{
Computes the effective sample size from historical data given a
discounting parameter.
}
\examples{
# With 100 historical observations and 50\% discounting
effective_sample_size(100, a0 = 0.5)  # Returns 50

}
