% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\docType{package}
\name{powerprior-package}
\alias{powerprior}
\alias{powerprior-package}
\title{Utility Functions for powerprior Package}
\description{
Additional helper functions for power prior analysis
}
\author{
\strong{Maintainer}: Yusuke Yamaguchi \email{yamagubed@gmail.com}

Authors:
\itemize{
  \item Yifei Huang
}

}
\keyword{internal}
