% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.shinyapp.R
\name{powerprior_launch_shinyapp}
\alias{powerprior_launch_shinyapp}
\title{Launch the Shiny App}
\usage{
powerprior_launch_shinyapp()
}
\value{
No return value, called for side effects (launches Shiny app)
}
\description{
This function launches the Shiny application.
}
