% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{print.compatibility_check}
\alias{print.compatibility_check}
\title{Print method for compatibility_check}
\usage{
\method{print}{compatibility_check}(x, digits = 4, ...)
}
\arguments{
\item{x}{Object of class "compatibility_check"}

\item{digits}{Number of digits to round to}

\item{...}{Additional arguments}
}
\value{
Invisibly returns the input object (for method chaining)
}
\description{
Print method for compatibility_check
}
