% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate.r
\name{print.posterior_multivariate}
\alias{print.posterior_multivariate}
\title{Print method for posterior_multivariate}
\usage{
\method{print}{posterior_multivariate}(x, ...)
}
\arguments{
\item{x}{Object of class "posterior_multivariate"}

\item{...}{Additional arguments}
}
\description{
Print method for posterior_multivariate
}
