% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.r
\name{print.posterior_univariate}
\alias{print.posterior_univariate}
\title{Print method for posterior_univariate}
\usage{
\method{print}{posterior_univariate}(x, ...)
}
\arguments{
\item{x}{Object of class "posterior_univariate"}

\item{...}{Additional arguments}
}
\description{
Print method for posterior_univariate
}
