% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{print.powerprior_comparison}
\alias{print.powerprior_comparison}
\title{Print method for powerprior_comparison}
\usage{
\method{print}{powerprior_comparison}(x, digits = 4, ...)
}
\arguments{
\item{x}{Object of class "powerprior_comparison"}

\item{digits}{Number of digits to round to}

\item{...}{Additional arguments}
}
\value{
Invisibly returns the input object (for method chaining)
}
\description{
Print method for powerprior_comparison
}
