% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate.r
\name{print.powerprior_multivariate}
\alias{print.powerprior_multivariate}
\title{Print method for powerprior_multivariate}
\usage{
\method{print}{powerprior_multivariate}(x, ...)
}
\arguments{
\item{x}{Object of class "powerprior_multivariate"}

\item{...}{Additional arguments}
}
\value{
Invisibly returns the input object (for method chaining)
}
\description{
Print method for powerprior_multivariate
}
