% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{print.powerprior_summary}
\alias{print.powerprior_summary}
\title{Print method for powerprior_summary}
\usage{
\method{print}{powerprior_summary}(x, digits = 4, ...)
}
\arguments{
\item{x}{Object of class "powerprior_summary"}

\item{digits}{Number of digits to round to}

\item{...}{Additional arguments}
}
\description{
Print method for powerprior_summary
}
