% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.r
\name{print.powerprior_univariate}
\alias{print.powerprior_univariate}
\title{Print method for powerprior_univariate}
\usage{
\method{print}{powerprior_univariate}(x, ...)
}
\arguments{
\item{x}{Object of class "powerprior_univariate"}

\item{...}{Additional arguments}
}
\description{
Print method for powerprior_univariate
}
